update redux store after creating new questionnaire
