+++ src/containers/questionnaire/QuestionnaireActions.js
@@ -3,6 +3,12 @@
 import { newRequestSequence } from 'redux-reqseq';
 import type { RequestSequence } from 'redux-reqseq';
 
+const CREATE_QUESTIONNAIRE :'CREATE_QUESTIONNAIRE' = 'CREATE_QUESTIONNAIRE';
+const createQuestionnaire :RequestSequence = newRequestSequence(CREATE_QUESTIONNAIRE);
+
+const DELETE_QUESTIONNAIRE :'DELETE_QUESTIONNAIRE' = 'DELETE_QUESTIONNAIRE';
+const deleteQuestionnaire :RequestSequence = newRequestSequence(DELETE_QUESTIONNAIRE);
+
 const DOWNLOAD_QUESTIONNAIRE_RESPONSES :'DOWNLOAD_QUESTIONNAIRE_RESPONSES' = 'DOWNLOAD_QUESTIONNAIRE_RESPONSES';
 const downloadQuestionnaireResponses :RequestSequence = newRequestSequence(DOWNLOAD_QUESTIONNAIRE_RESPONSES);
 
@@ -19,11 +25,15 @@ const SUBMIT_QUESTIONNAIRE :'SUBMIT_QUESTIONNAIRE' = 'SUBMIT_QUESTIONNAIRE';
 const submitQuestionnaire :RequestSequence = newRequestSequence(SUBMIT_QUESTIONNAIRE);
 
 export {
+  CREATE_QUESTIONNAIRE,
+  DELETE_QUESTIONNAIRE,
   DOWNLOAD_QUESTIONNAIRE_RESPONSES,
   GET_QUESTIONNAIRE,
   GET_QUESTIONNAIRE_RESPONSES,
   GET_STUDY_QUESTIONNAIRES,
   SUBMIT_QUESTIONNAIRE,
+  createQuestionnaire,
+  deleteQuestionnaire,
   downloadQuestionnaireResponses,
   getQuestionnaire,
   getQuestionnaireResponses,
