+++ src/containers/questionnaire/QuestionnaireReducer.js
@@ -2,26 +2,32 @@
 
 import {
   Map,
-  fromJS
+  fromJS,
+  getIn
 } from 'immutable';
+import { Constants } from 'lattice';
 import { RequestStates } from 'redux-reqseq';
 import type { SequenceAction } from 'redux-reqseq';
 
 import {
+  CREATE_QUESTIONNAIRE,
   DOWNLOAD_QUESTIONNAIRE_RESPONSES,
   GET_QUESTIONNAIRE,
   GET_QUESTIONNAIRE_RESPONSES,
   GET_STUDY_QUESTIONNAIRES,
   SUBMIT_QUESTIONNAIRE,
+  createQuestionnaire,
   downloadQuestionnaireResponses,
   getQuestionnaire,
   getQuestionnaireResponses,
   getStudyQuestionnaires,
-  submitQuestionnaire,
+  submitQuestionnaire
 } from './QuestionnaireActions';
 
 import { RESET_REQUEST_STATE } from '../../core/redux/ReduxActions';
 import { QUESTIONNAIRE_REDUX_CONSTANTS } from '../../utils/constants/ReduxConstants';
+
+const { OPENLATTICE_ID_FQN } = Constants;
 
 const {
   ANSWER_QUESTION_ID_MAP,
@@ -33,6 +39,9 @@ const {
 } = QUESTIONNAIRE_REDUX_CONSTANTS;
 
 const INITIAL_STATE = fromJS({
+  [CREATE_QUESTIONNAIRE]: {
+    requestState: RequestStates.STANDBY
+  },
   [DOWNLOAD_QUESTIONNAIRE_RESPONSES]: {
     requestState: RequestStates.STANDBY
   },
@@ -131,6 +140,31 @@ export default function questionnareReducer(state :Map = INITIAL_STATE, action :
         REQUEST: () => state.setIn([SUBMIT_QUESTIONNAIRE, 'requestState'], RequestStates.PENDING),
         FAILURE: () => state.setIn([SUBMIT_QUESTIONNAIRE, 'requestState'], RequestStates.FAILURE),
         SUCCESS: () => state.setIn([SUBMIT_QUESTIONNAIRE, 'requestState'], RequestStates.SUCCESS)
+      });
+    }
+
+    case createQuestionnaire.case(action.type): {
+      return createQuestionnaire.reducer(state, action, {
+        REQUEST: () => state.setIn([CREATE_QUESTIONNAIRE, 'requestState'], RequestStates.PENDING),
+        FAILURE: () => state.setIn([CREATE_QUESTIONNAIRE, 'requestState'], RequestStates.FAILURE),
+        SUCCESS: () => {
+          const {
+            questionEntities,
+            questionnaireEntity,
+            studyEKID
+          } = action.value;
+
+          const questionnaireEKID = getIn(questionnaireEntity, [OPENLATTICE_ID_FQN, 0]);
+
+          // console.log(questionEntities);
+          // console.log(questionnaireEntity);
+          // console.log(studyEKID);
+
+          return state
+            .setIn([QUESTIONNAIRE_QUESTIONS, questionnaireEKID], fromJS(questionEntities))
+            .setIn([STUDY_QUESTIONNAIRES, studyEKID, questionnaireEKID], fromJS(questionnaireEntity))
+            .setIn([CREATE_QUESTIONNAIRE, 'requestState'], RequestStates.SUCCESS);
+        }
       });
     }
     default:
